requirements.txt
streamlit>=1.20
pandas
Pillow
opencv-python-headless
face-recognition
